seirl/swh-dedup/requirements.txt
borgbackup
pyrabin
python-magic
